@@ -1,4 +1,5 @@
 using System;
+using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Options;
 
@@ -15,7 +16,8 @@ namespace BaGet.Core
             _apiKey = string.IsNullOrEmpty(options.Value.ApiKey) ? null : options.Value.ApiKey;
         }
 
-        public Task<bool> AuthenticateAsync(string apiKey) => Task.FromResult(Authenticate(apiKey));
+        public Task<bool> AuthenticateAsync(string apiKey, CancellationToken cancellationToken)
+            => Task.FromResult(Authenticate(apiKey));
 
         private bool Authenticate(string apiKey)
         {
