@@ -1,3 +1,4 @@
+using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
@@ -22,6 +23,14 @@ namespace BaGet.Core
         /// </summary>
         bool SupportsLimitInSubqueries { get; }
 
-        Task<int> SaveChangesAsync();
+        Task<int> SaveChangesAsync(CancellationToken cancellationToken);
+
+        /// <summary>
+        /// Applies any pending migrations for the context to the database.
+        /// Creates the database if it does not already exist.
+        /// </summary>
+        /// <param name="cancellationToken">A token to cancel the task.</param>
+        /// <returns>A task that completes once migrations are applied.</returns>
+        Task RunMigrationsAsync(CancellationToken cancellationToken);
     }
 }
