@@ -14,79 +14,111 @@ namespace BaGet.Database.MySql.Migrations
         {
 #pragma warning disable 612, 618
             modelBuilder
-                .HasAnnotation("ProductVersion", "2.2.6-servicing-10079")
+                .HasAnnotation("ProductVersion", "3.1.1")
                 .HasAnnotation("Relational:MaxIdentifierLength", 64);
 
-            modelBuilder.Entity("BaGet.Core.Entities.Package", b =>
+            modelBuilder.Entity("BaGet.Core.Package", b =>
                 {
                     b.Property<int>("Key")
-                        .ValueGeneratedOnAdd();
+                        .ValueGeneratedOnAdd()
+                        .HasColumnType("int");
 
                     b.Property<string>("Authors")
+                        .HasColumnType("longtext CHARACTER SET utf8mb4")
                         .HasMaxLength(4000);
 
                     b.Property<string>("Description")
+                        .HasColumnType("longtext CHARACTER SET utf8mb4")
                         .HasMaxLength(4000);
 
-                    b.Property<long>("Downloads");
+                    b.Property<long>("Downloads")
+                        .HasColumnType("bigint");
 
-                    b.Property<bool>("HasReadme");
+                    b.Property<bool>("HasEmbeddedIcon")
+                        .HasColumnType("tinyint(1)");
+
+                    b.Property<bool>("HasReadme")
+                        .HasColumnType("tinyint(1)");
 
                     b.Property<string>("IconUrl")
+                        .HasColumnType("longtext CHARACTER SET utf8mb4")
                         .HasMaxLength(4000);
 
                     b.Property<string>("Id")
                         .IsRequired()
+                        .HasColumnType("varchar(128) CHARACTER SET utf8mb4")
                         .HasMaxLength(128);
 
-                    b.Property<bool>("IsPrerelease");
+                    b.Property<bool>("IsPrerelease")
+                        .HasColumnType("tinyint(1)");
 
                     b.Property<string>("Language")
+                        .HasColumnType("varchar(20) CHARACTER SET utf8mb4")
                         .HasMaxLength(20);
 
                     b.Property<string>("LicenseUrl")
+                        .HasColumnType("longtext CHARACTER SET utf8mb4")
                         .HasMaxLength(4000);
 
-                    b.Property<bool>("Listed");
+                    b.Property<bool>("Listed")
+                        .HasColumnType("tinyint(1)");
 
                     b.Property<string>("MinClientVersion")
+                        .HasColumnType("varchar(44) CHARACTER SET utf8mb4")
                         .HasMaxLength(44);
 
                     b.Property<string>("NormalizedVersionString")
                         .IsRequired()
                         .HasColumnName("Version")
+                        .HasColumnType("varchar(64) CHARACTER SET utf8mb4")
                         .HasMaxLength(64);
 
                     b.Property<string>("OriginalVersionString")
                         .HasColumnName("OriginalVersion")
+                        .HasColumnType("varchar(64) CHARACTER SET utf8mb4")
                         .HasMaxLength(64);
 
                     b.Property<string>("ProjectUrl")
+                        .HasColumnType("longtext CHARACTER SET utf8mb4")
                         .HasMaxLength(4000);
 
-                    b.Property<DateTime>("Published");
+                    b.Property<DateTime>("Published")
+                        .HasColumnType("datetime(6)");
+
+                    b.Property<string>("ReleaseNotes")
+                        .HasColumnName("ReleaseNotes")
+                        .HasColumnType("longtext CHARACTER SET utf8mb4")
+                        .HasMaxLength(4000);
 
                     b.Property<string>("RepositoryType")
+                        .HasColumnType("varchar(100) CHARACTER SET utf8mb4")
                         .HasMaxLength(100);
 
                     b.Property<string>("RepositoryUrl")
+                        .HasColumnType("longtext CHARACTER SET utf8mb4")
                         .HasMaxLength(4000);
 
-                    b.Property<bool>("RequireLicenseAcceptance");
+                    b.Property<bool>("RequireLicenseAcceptance")
+                        .HasColumnType("tinyint(1)");
 
                     b.Property<DateTime?>("RowVersion")
                         .IsConcurrencyToken()
-                        .ValueGeneratedOnAddOrUpdate();
+                        .ValueGeneratedOnAddOrUpdate()
+                        .HasColumnType("timestamp(6)");
 
-                    b.Property<int>("SemVerLevel");
+                    b.Property<int>("SemVerLevel")
+                        .HasColumnType("int");
 
                     b.Property<string>("Summary")
+                        .HasColumnType("longtext CHARACTER SET utf8mb4")
                         .HasMaxLength(4000);
 
                     b.Property<string>("Tags")
+                        .HasColumnType("longtext CHARACTER SET utf8mb4")
                         .HasMaxLength(4000);
 
                     b.Property<string>("Title")
+                        .HasColumnType("varchar(256) CHARACTER SET utf8mb4")
                         .HasMaxLength(256);
 
                     b.HasKey("Key");
@@ -99,20 +131,25 @@ namespace BaGet.Database.MySql.Migrations
                     b.ToTable("Packages");
                 });
 
-            modelBuilder.Entity("BaGet.Core.Entities.PackageDependency", b =>
+            modelBuilder.Entity("BaGet.Core.PackageDependency", b =>
                 {
                     b.Property<int>("Key")
-                        .ValueGeneratedOnAdd();
+                        .ValueGeneratedOnAdd()
+                        .HasColumnType("int");
 
                     b.Property<string>("Id")
+                        .HasColumnType("varchar(128) CHARACTER SET utf8mb4")
                         .HasMaxLength(128);
 
-                    b.Property<int?>("PackageKey");
+                    b.Property<int?>("PackageKey")
+                        .HasColumnType("int");
 
                     b.Property<string>("TargetFramework")
+                        .HasColumnType("varchar(256) CHARACTER SET utf8mb4")
                         .HasMaxLength(256);
 
                     b.Property<string>("VersionRange")
+                        .HasColumnType("varchar(256) CHARACTER SET utf8mb4")
                         .HasMaxLength(256);
 
                     b.HasKey("Key");
@@ -124,17 +161,21 @@ namespace BaGet.Database.MySql.Migrations
                     b.ToTable("PackageDependencies");
                 });
 
-            modelBuilder.Entity("BaGet.Core.Entities.PackageType", b =>
+            modelBuilder.Entity("BaGet.Core.PackageType", b =>
                 {
                     b.Property<int>("Key")
-                        .ValueGeneratedOnAdd();
+                        .ValueGeneratedOnAdd()
+                        .HasColumnType("int");
 
                     b.Property<string>("Name")
+                        .HasColumnType("varchar(512) CHARACTER SET utf8mb4")
                         .HasMaxLength(512);
 
-                    b.Property<int>("PackageKey");
+                    b.Property<int>("PackageKey")
+                        .HasColumnType("int");
 
                     b.Property<string>("Version")
+                        .HasColumnType("varchar(64) CHARACTER SET utf8mb4")
                         .HasMaxLength(64);
 
                     b.HasKey("Key");
@@ -146,15 +187,18 @@ namespace BaGet.Database.MySql.Migrations
                     b.ToTable("PackageTypes");
                 });
 
-            modelBuilder.Entity("BaGet.Core.Entities.TargetFramework", b =>
+            modelBuilder.Entity("BaGet.Core.TargetFramework", b =>
                 {
                     b.Property<int>("Key")
-                        .ValueGeneratedOnAdd();
+                        .ValueGeneratedOnAdd()
+                        .HasColumnType("int");
 
                     b.Property<string>("Moniker")
+                        .HasColumnType("varchar(256) CHARACTER SET utf8mb4")
                         .HasMaxLength(256);
 
-                    b.Property<int>("PackageKey");
+                    b.Property<int>("PackageKey")
+                        .HasColumnType("int");
 
                     b.HasKey("Key");
 
@@ -165,27 +209,29 @@ namespace BaGet.Database.MySql.Migrations
                     b.ToTable("TargetFrameworks");
                 });
 
-            modelBuilder.Entity("BaGet.Core.Entities.PackageDependency", b =>
+            modelBuilder.Entity("BaGet.Core.PackageDependency", b =>
                 {
-                    b.HasOne("BaGet.Core.Entities.Package", "Package")
+                    b.HasOne("BaGet.Core.Package", "Package")
                         .WithMany("Dependencies")
                         .HasForeignKey("PackageKey");
                 });
 
-            modelBuilder.Entity("BaGet.Core.Entities.PackageType", b =>
+            modelBuilder.Entity("BaGet.Core.PackageType", b =>
                 {
-                    b.HasOne("BaGet.Core.Entities.Package", "Package")
+                    b.HasOne("BaGet.Core.Package", "Package")
                         .WithMany("PackageTypes")
                         .HasForeignKey("PackageKey")
-                        .OnDelete(DeleteBehavior.Cascade);
+                        .OnDelete(DeleteBehavior.Cascade)
+                        .IsRequired();
                 });
 
-            modelBuilder.Entity("BaGet.Core.Entities.TargetFramework", b =>
+            modelBuilder.Entity("BaGet.Core.TargetFramework", b =>
                 {
-                    b.HasOne("BaGet.Core.Entities.Package", "Package")
+                    b.HasOne("BaGet.Core.Package", "Package")
                         .WithMany("TargetFrameworks")
                         .HasForeignKey("PackageKey")
-                        .OnDelete(DeleteBehavior.Cascade);
+                        .OnDelete(DeleteBehavior.Cascade)
+                        .IsRequired();
                 });
 #pragma warning restore 612, 618
         }
