@@ -15,81 +15,113 @@ namespace BaGet.Database.SqlServer.Migrations
         {
 #pragma warning disable 612, 618
             modelBuilder
-                .HasAnnotation("ProductVersion", "2.2.6-servicing-10079")
+                .HasAnnotation("ProductVersion", "3.1.1")
                 .HasAnnotation("Relational:MaxIdentifierLength", 128)
                 .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
-            modelBuilder.Entity("BaGet.Core.Entities.Package", b =>
+            modelBuilder.Entity("BaGet.Core.Package", b =>
                 {
                     b.Property<int>("Key")
                         .ValueGeneratedOnAdd()
+                        .HasColumnType("int")
                         .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
                     b.Property<string>("Authors")
+                        .HasColumnType("nvarchar(4000)")
                         .HasMaxLength(4000);
 
                     b.Property<string>("Description")
+                        .HasColumnType("nvarchar(4000)")
                         .HasMaxLength(4000);
 
-                    b.Property<long>("Downloads");
+                    b.Property<long>("Downloads")
+                        .HasColumnType("bigint");
 
-                    b.Property<bool>("HasReadme");
+                    b.Property<bool>("HasEmbeddedIcon")
+                        .HasColumnType("bit");
+
+                    b.Property<bool>("HasReadme")
+                        .HasColumnType("bit");
 
                     b.Property<string>("IconUrl")
+                        .HasColumnType("nvarchar(4000)")
                         .HasMaxLength(4000);
 
                     b.Property<string>("Id")
                         .IsRequired()
+                        .HasColumnType("nvarchar(128)")
                         .HasMaxLength(128);
 
-                    b.Property<bool>("IsPrerelease");
+                    b.Property<bool>("IsPrerelease")
+                        .HasColumnType("bit");
 
                     b.Property<string>("Language")
+                        .HasColumnType("nvarchar(20)")
                         .HasMaxLength(20);
 
                     b.Property<string>("LicenseUrl")
+                        .HasColumnType("nvarchar(4000)")
                         .HasMaxLength(4000);
 
-                    b.Property<bool>("Listed");
+                    b.Property<bool>("Listed")
+                        .HasColumnType("bit");
 
                     b.Property<string>("MinClientVersion")
+                        .HasColumnType("nvarchar(44)")
                         .HasMaxLength(44);
 
                     b.Property<string>("NormalizedVersionString")
                         .IsRequired()
                         .HasColumnName("Version")
+                        .HasColumnType("nvarchar(64)")
                         .HasMaxLength(64);
 
                     b.Property<string>("OriginalVersionString")
                         .HasColumnName("OriginalVersion")
+                        .HasColumnType("nvarchar(64)")
                         .HasMaxLength(64);
 
                     b.Property<string>("ProjectUrl")
+                        .HasColumnType("nvarchar(4000)")
                         .HasMaxLength(4000);
 
-                    b.Property<DateTime>("Published");
+                    b.Property<DateTime>("Published")
+                        .HasColumnType("datetime2");
+
+                    b.Property<string>("ReleaseNotes")
+                        .HasColumnName("ReleaseNotes")
+                        .HasColumnType("nvarchar(4000)")
+                        .HasMaxLength(4000);
 
                     b.Property<string>("RepositoryType")
+                        .HasColumnType("nvarchar(100)")
                         .HasMaxLength(100);
 
                     b.Property<string>("RepositoryUrl")
+                        .HasColumnType("nvarchar(4000)")
                         .HasMaxLength(4000);
 
-                    b.Property<bool>("RequireLicenseAcceptance");
+                    b.Property<bool>("RequireLicenseAcceptance")
+                        .HasColumnType("bit");
 
                     b.Property<byte[]>("RowVersion")
                         .IsConcurrencyToken()
-                        .ValueGeneratedOnAddOrUpdate();
+                        .ValueGeneratedOnAddOrUpdate()
+                        .HasColumnType("rowversion");
 
-                    b.Property<int>("SemVerLevel");
+                    b.Property<int>("SemVerLevel")
+                        .HasColumnType("int");
 
                     b.Property<string>("Summary")
+                        .HasColumnType("nvarchar(4000)")
                         .HasMaxLength(4000);
 
                     b.Property<string>("Tags")
+                        .HasColumnType("nvarchar(4000)")
                         .HasMaxLength(4000);
 
                     b.Property<string>("Title")
+                        .HasColumnType("nvarchar(256)")
                         .HasMaxLength(256);
 
                     b.HasKey("Key");
@@ -102,21 +134,26 @@ namespace BaGet.Database.SqlServer.Migrations
                     b.ToTable("Packages");
                 });
 
-            modelBuilder.Entity("BaGet.Core.Entities.PackageDependency", b =>
+            modelBuilder.Entity("BaGet.Core.PackageDependency", b =>
                 {
                     b.Property<int>("Key")
                         .ValueGeneratedOnAdd()
+                        .HasColumnType("int")
                         .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
                     b.Property<string>("Id")
+                        .HasColumnType("nvarchar(128)")
                         .HasMaxLength(128);
 
-                    b.Property<int?>("PackageKey");
+                    b.Property<int?>("PackageKey")
+                        .HasColumnType("int");
 
                     b.Property<string>("TargetFramework")
+                        .HasColumnType("nvarchar(256)")
                         .HasMaxLength(256);
 
                     b.Property<string>("VersionRange")
+                        .HasColumnType("nvarchar(256)")
                         .HasMaxLength(256);
 
                     b.HasKey("Key");
@@ -128,18 +165,22 @@ namespace BaGet.Database.SqlServer.Migrations
                     b.ToTable("PackageDependencies");
                 });
 
-            modelBuilder.Entity("BaGet.Core.Entities.PackageType", b =>
+            modelBuilder.Entity("BaGet.Core.PackageType", b =>
                 {
                     b.Property<int>("Key")
                         .ValueGeneratedOnAdd()
+                        .HasColumnType("int")
                         .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
                     b.Property<string>("Name")
+                        .HasColumnType("nvarchar(512)")
                         .HasMaxLength(512);
 
-                    b.Property<int>("PackageKey");
+                    b.Property<int>("PackageKey")
+                        .HasColumnType("int");
 
                     b.Property<string>("Version")
+                        .HasColumnType("nvarchar(64)")
                         .HasMaxLength(64);
 
                     b.HasKey("Key");
@@ -151,16 +192,19 @@ namespace BaGet.Database.SqlServer.Migrations
                     b.ToTable("PackageTypes");
                 });
 
-            modelBuilder.Entity("BaGet.Core.Entities.TargetFramework", b =>
+            modelBuilder.Entity("BaGet.Core.TargetFramework", b =>
                 {
                     b.Property<int>("Key")
                         .ValueGeneratedOnAdd()
+                        .HasColumnType("int")
                         .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
                     b.Property<string>("Moniker")
+                        .HasColumnType("nvarchar(256)")
                         .HasMaxLength(256);
 
-                    b.Property<int>("PackageKey");
+                    b.Property<int>("PackageKey")
+                        .HasColumnType("int");
 
                     b.HasKey("Key");
 
@@ -171,27 +215,29 @@ namespace BaGet.Database.SqlServer.Migrations
                     b.ToTable("TargetFrameworks");
                 });
 
-            modelBuilder.Entity("BaGet.Core.Entities.PackageDependency", b =>
+            modelBuilder.Entity("BaGet.Core.PackageDependency", b =>
                 {
-                    b.HasOne("BaGet.Core.Entities.Package", "Package")
+                    b.HasOne("BaGet.Core.Package", "Package")
                         .WithMany("Dependencies")
                         .HasForeignKey("PackageKey");
                 });
 
-            modelBuilder.Entity("BaGet.Core.Entities.PackageType", b =>
+            modelBuilder.Entity("BaGet.Core.PackageType", b =>
                 {
-                    b.HasOne("BaGet.Core.Entities.Package", "Package")
+                    b.HasOne("BaGet.Core.Package", "Package")
                         .WithMany("PackageTypes")
                         .HasForeignKey("PackageKey")
-                        .OnDelete(DeleteBehavior.Cascade);
+                        .OnDelete(DeleteBehavior.Cascade)
+                        .IsRequired();
                 });
 
-            modelBuilder.Entity("BaGet.Core.Entities.TargetFramework", b =>
+            modelBuilder.Entity("BaGet.Core.TargetFramework", b =>
                 {
-                    b.HasOne("BaGet.Core.Entities.Package", "Package")
+                    b.HasOne("BaGet.Core.Package", "Package")
                         .WithMany("TargetFrameworks")
                         .HasForeignKey("PackageKey")
-                        .OnDelete(DeleteBehavior.Cascade);
+                        .OnDelete(DeleteBehavior.Cascade)
+                        .IsRequired();
                 });
 #pragma warning restore 612, 618
         }
