@@ -10,14 +10,14 @@ namespace BaGet.Core.Tests.Services
     {
         private readonly Mock<IPackageService> _packages;
         private readonly Mock<IPackageStorageService> _storage;
-        private readonly Mock<ISearchService> _search;
+        private readonly Mock<ISearchIndexer> _search;
         private readonly PackageIndexingService _target;
 
         public PackageIndexingServiceTests()
         {
             _packages = new Mock<IPackageService>();
             _storage = new Mock<IPackageStorageService>();
-            _search = new Mock<ISearchService>();
+            _search = new Mock<ISearchIndexer>();
 
             _target = new PackageIndexingService(
                 _packages.Object,
